feat(blend): export tooltip/popover/modal bases and types from node
Equivalence-preserving restructures: TooltipBaseProps, PopoverBaseProps
and ModalBaseProps carry only the already-platform-neutral props (web
props types are unchanged intersections). The popover enums + base and
the modal base live in new leaf modules (popoverV2.base.types.ts,
modalV2.base.types.ts) re-exported by the original types files, because
those files pull the ButtonV2/Skeleton barrels whose graphs reach
DOM-typed runtime the React-free node entry must never see. node.ts adds
the three token types, the tooltip/popover enums, and the bases. Dist
A/B: only node.js and chunk hashes shifted; TooltipV2/PopoverV2/ModalV2
suites pass and dist/node.js stays React-free.
